change price field from to NUMERIC from p Produfrom Products

+++ models/product.js
@@ -15,7 +15,7 @@ module.exports = (sequelize, DataTypes) => {
   }
   Product.init({
     name: DataTypes.STRING,
-    price: DataTypes.FLOAT,
+    price: DataTypes.NUMERIC,
     CategoryId: DataTypes.INTEGER
   }, {
     sequelize,
